6 motors in 90 to 0 degrees and to 90 degree
you will find a video and the code also the circuit to controlling 5 servo motors,  but i add more one first will see how the motors will change from 90 degrees to 0 then i change the code to make it in a 90-degree position. this is for the third training program with a smart method company.