task3 r-charger battery 3 models smart-method training program
The first one:
In this circuit, the re-charge will be in one ore full re-charger, by giving a 12 volt and 5
amperes for a 12volt battery.
First the transformer and after it, the rectifier bridge for converting the current from AC to
DC came from the supplier, after that the capacitor will smoothing in a low value like C1, the
LED will be our leader to know the re-charger with an Ampere meter to see the value from
the positive rail, so when the voltage been more than 13.8 volts the current will stop flows,
and the battery will be full, and the reading of the meter will be zero.
The second one:
In this circuit, the car lamp will be lighting if the battery is re-charging by seeing the
brightness on it, and if it is full the lamp will turn off, but if the lamp still fully bright for
more than 30 min that will mean the battery is dead.
Here the transformer step-down from a 15--0-15 volt and 2 Ampere will transfer the
charger current from AC source and the 2 diodes will rectifier 3amper current which they
can handle it, also the capacitor will filtering in low-value to do some ripples for better
charging to the battery in a better way. So the trick of this circuit depending on the car lamp
the bulb is dependent on the flow current to the battery, so it will be light if the battery
needs to re-charge too if it had some charger so until it charges the light will lighting till be
closer from full charge you will see the brightness low, also the lamp will resist the flow of
current as a resistor.
The third one:
In this circuit will find a re-charger battery that can handle higher currents, having two
diodes one will light when the recharger start and the other when it will be full.
First, the transformer can transform the voltage from 220 to 14-20 volt, then the bridge will
handling the current as it likes to handle to transform the AC current to DC and smoothing it
make it stronger.
Ziner diode is 12 volt when the voltage on both sides of the battery been more than 13-volt
Ziner will open a pass for current to the Base part transistor for it to the relay to stop the recharge and you will know it by the green LED light. In addition, the resistors for limiting the
current through the LEDs, transistor, and the Ziner diode.
for these circuits, I will use a NiMH or Nickel Metal Hydride battery or lithium, and my
reason for these choices is because the size of both of them have a small size also not that
much heavier if we put it in our robot also they are cheaper but there is kind of lithium
battery are expensive
In each of them, we need to take care to recharge them the first one are slow if we
recharger it, but the other one fast but I prefer the third circuit for it with using the
cheaper one of lithium, not the expensive one it looks like the battery in or cell phones,
but the less price one like the normal battery games in shape but of course it a higher price
.more than the NiMH. And for me, I will use the lithium one.