task 4 Robot eyes system Smart method training program
There are two ways to connect the LED eyes in parallel and serial, the parallel contact the voltage source will be equally in both LEDs means they will lite in the same lightness way but if one of them damaged the other one will be ok, unlike then serial the voltage source will division to two-part and if one led been damaged the other will be also damaged too.
for the serial circuit, we need the power a little more for light to be equal for both LEDs, but for parallel, we do not need that because it will divide. 
I use a PIR sensor for saving power, which means if any things moving in the PIR area sensing the LED will turn on.
in this circuit you will see the re-charger system for these eyes, I use a potentiometer because in the circuit wizard program there is no PIR sensor, it works in the same way as the PIR sensor.