task7 with smart methode
In this task, i use an esp32 with two led and the idea is to light the led from a website after I made the esp access point I will use the IP which seen in serial monitor then changes my wifi to the ESP wifi, next write the password, after it, it will take me to the web page with to button tow light the led, to press it then the light will on.  
i actually tried with ESP12E but I got some trouble with it until it damage. so I read the code and change it to make it usable with ESP32.